agent_03_instagram_crawler/requirements.txt
playwright==1.48.0
python-dotenv==1.0.1
pymysql==1.1.1